cmd: move updatePolicyKey to helpers.go
This function is used both by `cilium bpf policy add` and `cilium bpf policy
delete`,so move it to the file where helpers that are common to more than
one command are located.

Signed-off by: Ian Vernon <ian@cilium.io>
